Setup logging and its configurations
Add logging

Add a logging.conf file to handle two loggers and their configurations:
root and __main__ (the application package)
The root has to be declared in the logging configuration as part of python logging package.

Set the logging.fileConfig as the logging.conf file

Add a stream handler and a basic formatter to the loggers for root and __main__ packages.

Print logs to stdout