Add tests for laa_references POST calls
Create a folder laa_references in the tests/routers directory for readability.

Add a post request tests for laa_references

Pull out the mocked responses of the POST request to CDA in each test.
This is because the only difference between the cda request calls for a POST laa_references is the request body and not the endpoint.
This meant that tests were failing because it wasn't matching the correct mock route.

Update the fixtures by adding a POST laa_references route and updating the PATCH route