Increase the httpx timeout to 60s
This is because data fetching from CDA can be slow. We get ReadTimeouts at 15s when there are requests that take just over 30s
I believe some of the 424 failed dependencies returned to CDUI is because of the read timeout maxing at 15seconds.

Increasing this to 60s should give us more time to wait for a response from laa-court-data-adaptor.